+*：修改包厢列表参数接收不上，支付回调异常处理。

+++ mazongjian-module-member/mazongjian-module-member-biz/src/main/java/com/yanzu/module/member/controller/app/pkg/vo/AppAdminPkgSaveReqVO.java
@@ -38,7 +38,7 @@ public class AppAdminPkgSaveReqVO {
     private String ksId;
 
     @Schema(description = "按包厢设置")
-    private List<Long> enableRoom;
+    private List<Integer> enableRoom;
 
     @Schema(description = "可用时间 0-23数字，空表示不限")
     private List<Integer> enableTime;
